+++ liquidity_pool_router/src/testutils.rs
@@ -1,5 +1,5 @@
 #![cfg(any(test, feature = "testutils"))]
-use soroban_sdk::{AccountId, BigInt, BytesN, Env};
+use soroban_sdk::{AccountId, BytesN, Env};
 
 use crate::LiquidityPoolRouterClient;
 
@@ -30,10 +30,10 @@ impl LiquidityPoolRouter {
         to: &AccountId,
         token_a: &BytesN<32>,
         token_b: &BytesN<32>,
-        desired_a: &BigInt,
-        min_a: &BigInt,
-        desired_b: &BigInt,
-        min_b: &BigInt,
+        desired_a: &i128,
+        min_a: &i128,
+        desired_b: &i128,
+        min_b: &i128,
     ) {
         self.client().with_source_account(&to).sf_deposit(
             liquidity_pool_wasm_hash,
@@ -51,8 +51,8 @@ impl LiquidityPoolRouter {
         to: &AccountId,
         sell: &BytesN<32>,
         buy: &BytesN<32>,
-        out: &BigInt,
-        in_max: &BigInt,
+        out: &i128,
+        in_max: &i128,
     ) {
         self.client()
             .with_source_account(&to)
@@ -64,9 +64,9 @@ impl LiquidityPoolRouter {
         to: &AccountId,
         token_a: &BytesN<32>,
         token_b: &BytesN<32>,
-        share_amount: &BigInt,
-        min_a: &BigInt,
-        min_b: &BigInt,
+        share_amount: &i128,
+        min_a: &i128,
+        min_b: &i128,
     ) {
         self.client().with_source_account(&to).sf_withdrw(
             token_a,
